MDPmotorcycles/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
beautifulsoup4
requests
selenium
webdriver-manager
